Split up chains
