+++ src/tests/category/chain/sunken_test.py
@@ -1,128 +1,15 @@
-from typing import Iterable, Set
+from typing import Set
 import pytest
-from modlunky2.mem.entities import EntityDBEntry, EntityType, Player
 
+from modlunky2.category.chain.common import ChainStatus, ChainStepEvaluator
+from modlunky2.category.chain.sunken import AbzuChain, DuatChain, SunkenChain
+from modlunky2.category.chain.testing import make_player_with_hh_items
+from modlunky2.mem.entities import EntityType, Player
 from modlunky2.mem.state import Items, Screen, State, Theme, WinState
-from modlunky2.ui.trackers.chain import (
-    AbzuChain,
-    ChainMixin,
-    ChainStatus,
-    ChainStepEvaluator,
-    ChainStepResult,
-    ChainStepper,
-    CommonSunkenChain,
-    CosmicOceanChain,
-    DuatChain,
-)
 from modlunky2.mem.testing import EntityMapBuilder
 
 
-def make_player_with_hh_items(
-    entity_map: EntityMapBuilder, hh_item_types: Iterable[EntityType]
-):
-    hh_item_ids = entity_map.add_trivial_entities(hh_item_types)
-    hh_id = entity_map.add_entity(Player(items=hh_item_ids))
-
-    return Player(linked_companion_child=hh_id)
-
-
-def make_player_with_hh_type(entity_map: EntityMapBuilder, hh_type: EntityType):
-    hh_entity_db = EntityDBEntry(id=hh_type)
-    hh_entity = Player(type=hh_entity_db)
-    hh_id = entity_map.add_entity(hh_entity)
-
-    return Player(linked_companion_child=hh_id)
-
-
-def fake_chain_step():
-    pass
-
-
-@pytest.mark.parametrize(
-    "status,next_step,msg_pattern",
-    [
-        (ChainStatus.UNSTARTED, fake_chain_step, "next_step to be None"),
-        (ChainStatus.IN_PROGRESS, None, "next_step to be set"),
-        (ChainStatus.IN_PROGRESS, dict(), "next_step to be callable"),
-        (ChainStatus.FAILED, fake_chain_step, "next_step to be None"),
-    ],
-)
-def test_chain_step_result_validation(status, next_step, msg_pattern):
-    with pytest.raises(ValueError, match=msg_pattern):
-        ChainStepResult(status, next_step)
-
-
-class TestChain(ChainMixin):
-    def still_unstarted(self, _unused1: State, _unused2: Set[EntityType]):
-        return self.unstarted()
-
-    def loopy_in_progress(self, _unused1: State, _unused2: Set[EntityType]):
-        return self.in_progress(self.loopy_in_progress)
-
-    def total_fail(self, _unused1: State, _unused2: Set[EntityType]):
-        return self.failed()
-
-    def step1_to_fail(self, _unused1: State, _unused2: Set[EntityType]):
-        return self.in_progress(self.step2_to_fail)
-
-    def step2_to_fail(self, _unused1: State, _unused2: Set[EntityType]):
-        return self.failed()
-
-    def step1_to_unstarted(self, _unused1: State, _unused2: Set[EntityType]):
-        return self.in_progress(self.step2_to_unstarted)
-
-    def step2_to_unstarted(self, _unused1: State, _unused2: Set[EntityType]):
-        return self.unstarted()
-
-
-@pytest.mark.parametrize(
-    "initial_step_name,expected_status",
-    [
-        ("still_unstarted", ChainStatus.UNSTARTED),
-        ("loopy_in_progress", ChainStatus.IN_PROGRESS),
-        ("total_fail", ChainStatus.FAILED),
-    ],
-)
-def test_chain_stepper_single(initial_step_name, expected_status):
-    test_chain = TestChain()
-    initial_step = getattr(test_chain, initial_step_name)
-    stepper = ChainStepper("test", initial_step)
-
-    status = stepper.evaluate(State(), set())
-    assert status == expected_status
-
-
-def test_chain_stepper_multi_progress():
-    test_chain = TestChain()
-    stepper = ChainStepper("test", test_chain.step1_to_fail)
-
-    status = stepper.evaluate(State(), set())
-    assert status == ChainStatus.IN_PROGRESS
-
-    status = stepper.evaluate(State(), set())
-    assert status == ChainStatus.FAILED
-
-    # Should still be failed
-    status = stepper.evaluate(State(), set())
-    assert status == ChainStatus.FAILED
-
-
-def test_chain_stepper_multi_unstarted():
-    test_chain = TestChain()
-    stepper = ChainStepper("test", test_chain.step1_to_unstarted)
-
-    status = stepper.evaluate(State(), set())
-    assert status == ChainStatus.IN_PROGRESS
-
-    status = stepper.evaluate(State(), set())
-    assert status == ChainStatus.UNSTARTED
-
-    # Should restart
-    status = stepper.evaluate(State(), set())
-    assert status == ChainStatus.IN_PROGRESS
-
-
-class MinimalCommonChain(CommonSunkenChain):
+class MinimalCommonChain(SunkenChain):
     world41_theme = Theme.TIDE_POOL
     world44_theme = Theme.ABZU
 
@@ -641,184 +528,6 @@ def test_keep_ankh(
 
     duat_chain = DuatChain()
     result = duat_chain.keep_ankh(game_state, player_item_set)
-
-    assert result.status == expected_status
-    if expected_status.in_progress:
-        assert result.next_step.__name__ == expected_step_name
-
-
-@pytest.mark.parametrize(
-    "world,player_item_set,expected_status,expected_step_name",
-    [
-        (1, set(), ChainStatus.UNSTARTED, None),
-        (2, set(), ChainStatus.UNSTARTED, None),
-        (2, {EntityType.ITEM_HOUYIBOW}, ChainStatus.IN_PROGRESS, "carry_bow_to_hundun"),
-        (3, set(), ChainStatus.FAILED, None),
-    ],
-)
-def test_collect_bow(world, player_item_set, expected_status, expected_step_name):
-    game_state = State(world=world)
-
-    co_chain = CosmicOceanChain()
-    result = co_chain.collect_bow(game_state, player_item_set)
-
-    assert result.status == expected_status
-    if expected_status.in_progress:
-        assert result.next_step.__name__ == expected_step_name
-
-
-@pytest.mark.parametrize(
-    "world,level,screen,win_state,player_item_set,hh_item_set,waddler_item_set,expected_status,expected_step_name",
-    [
-        (
-            2,
-            3,
-            Screen.LEVEL_TRANSITION,
-            WinState.NO_WIN,
-            {EntityType.ITEM_HOUYIBOW},
-            set(),  # hh
-            set(),  # waddler
-            ChainStatus.IN_PROGRESS,
-            "carry_bow_to_hundun",
-        ),
-        (
-            2,
-            3,
-            Screen.LEVEL_TRANSITION,
-            WinState.NO_WIN,
-            set(),  # player
-            set(),  # hh
-            set(),  # waddler
-            ChainStatus.FAILED,
-            None,
-        ),
-        (
-            2,
-            3,
-            Screen.LEVEL_TRANSITION,
-            WinState.NO_WIN,
-            set(),  # playere
-            {EntityType.ITEM_HOUYIBOW},  # hh
-            set(),  # waddler
-            ChainStatus.IN_PROGRESS,
-            "carry_bow_to_hundun",
-        ),
-        (
-            3,
-            1,
-            Screen.LEVEL_TRANSITION,
-            WinState.NO_WIN,
-            set(),  # player
-            set(),  # hh
-            {EntityType.ITEM_HOUYIBOW},  # waddler
-            ChainStatus.IN_PROGRESS,
-            "carry_bow_to_hundun",
-        ),
-        # Missed Waddler
-        (
-            7,
-            2,
-            Screen.LEVEL_TRANSITION,
-            WinState.NO_WIN,
-            set(),  # player
-            set(),  # hh
-            {EntityType.ITEM_HOUYIBOW},  # waddler
-            ChainStatus.FAILED,
-            None,
-        ),
-        (
-            7,
-            3,
-            Screen.LEVEL_TRANSITION,
-            WinState.NO_WIN,
-            {EntityType.ITEM_HOUYIBOW},
-            set(),  # hh
-            set(),  # waddler
-            ChainStatus.IN_PROGRESS,
-            "win_via_co",
-        ),
-        (
-            6,
-            4,
-            Screen.ENDING,
-            WinState.TIAMAT,
-            {EntityType.ITEM_HOUYIBOW},
-            set(),  # hh
-            set(),  # waddler
-            ChainStatus.FAILED,
-            None,
-        ),
-        (
-            7,
-            4,
-            Screen.ENDING,
-            WinState.HUNDUN,
-            {EntityType.ITEM_HOUYIBOW},
-            set(),  # hh
-            set(),  # waddler
-            ChainStatus.FAILED,
-            None,
-        ),
-        # Score screen is in base camp
-        (
-            1,
-            1,
-            Screen.SCORES,
-            WinState.TIAMAT,
-            {EntityType.ITEM_HOUYIBOW},
-            set(),  # hh
-            set(),  # waddler
-            ChainStatus.FAILED,
-            None,
-        ),
-    ],
-)
-def carry_bow_to_hundun(
-    world,
-    level,
-    screen,
-    win_state,
-    player_item_set,
-    hh_item_set,
-    waddler_item_set,
-    expected_status,
-    expected_step_name,
-):
-    entity_map = EntityMapBuilder()
-    player = make_player_with_hh_items(entity_map, hh_item_set)
-    game_state = State(
-        world=world,
-        level=level,
-        screen=screen,
-        win_state=win_state,
-        waddler_storage=tuple(waddler_item_set),
-        items=Items(players=(player,)),
-        instance_id_to_pointer=entity_map.build(),
-    )
-
-    co_chain = CosmicOceanChain()
-    result = co_chain.carry_bow_to_hundun(game_state, player_item_set)
-
-    assert result.status == expected_status
-    if expected_status.in_progress:
-        assert result.next_step.__name__ == expected_step_name
-
-
-@pytest.mark.parametrize(
-    "win_state,expected_status,expected_step_name",
-    [
-        (WinState.NO_WIN, ChainStatus.IN_PROGRESS, "win_via_co"),
-        (WinState.TIAMAT, ChainStatus.FAILED, None),
-        (WinState.HUNDUN, ChainStatus.FAILED, None),
-        (WinState.COSMIC_OCEAN, ChainStatus.IN_PROGRESS, "win_via_co"),
-    ],
-)
-def test_win_via_co(win_state, expected_status, expected_step_name):
-    game_state = State(win_state=win_state)
-    player_item_set = set()
-
-    co_chain = CosmicOceanChain()
-    result = co_chain.win_via_co(game_state, player_item_set)
 
     assert result.status == expected_status
     if expected_status.in_progress:
